Fix unit tests

+++ modules/tangos/tools/add_simulation.py
@@ -76,7 +76,7 @@ class SimulationAdderUpdater(object):
         n_tot = []
         for num, NDM, Nstar, Ngas in self.simulation_output.enumerate_halos(ts.extension):
             n_tot.append(NDM+Nstar+Ngas)
-        ids = np.zeros(len(n_tot))
+        ids = np.zeros(len(n_tot), dtype=int)
         ids[np.argsort(np.array(n_tot))[::-1]] = np.arange(len(n_tot)) + 1
         cnt = 1
         for num, NDM, Nstar, Ngas in self.simulation_output.enumerate_halos(ts.extension):
